+++ src/test/java/it/isislab/dmason/test/sim/field/grid/numeric/thin/DDoubleGrid2DXYThinTester.java
@@ -182,7 +182,7 @@ public class DDoubleGrid2DXYThinTester {
 	@Test
 	public void testSetDistributedObjectLocation() throws DMasonException {
 
-		for (int i = 0; i < numLoop; i++) {
+		for (double i = 0; i < numLoop; i++) {
 			Int2D location = toTest.getAvailableRandomLocation();
 			assertTrue(toTest.setDistributedObjectLocation(location, /* grid value */
 					i, /* SimState */ss));
@@ -197,7 +197,7 @@ public class DDoubleGrid2DXYThinTester {
 	public void testGetState() throws DMasonException {
 
 		// i'm moving an agent in the DistributedState
-		for (int i = 0; i < numLoop; i++) {
+		for (double i = 0; i < numLoop; i++) {
 			Int2D location = toTest.getAvailableRandomLocation();
 			toTest.setDistributedObjectLocation(location, /* grid value */
 					i, /* SimState */ss);
